Fixing DMA Bandwidth and Cleaning Up DMA for Release (#16)
- Increasing bandwidth to (nominal) 8.67GBps.
- Cleaning up SW app, check for errors and return values:
-1: Bandwidth requirement not met
   - 0: Success
   - 1-N: Number of errors incurred.
   - Renaming key modules to be more specific
- Adhering to coding guidelines by removing implicit port instantiations

+++ tutorial/afu_types/01_pim_ifc/dma/sw/dma_util.h
@@ -12,8 +12,8 @@ typedef enum dma_mode {
 } e_dma_mode;
 
 typedef struct __attribute__((__packed__))  {
-   uint32_t src_address;
-   uint32_t dest_address;
+   uint64_t src_address;
+   uint64_t dest_address;
    uint32_t len;
    uint32_t control;
 } dma_descriptor_t;
